--- a/Makefile
+++ b/Makefile
@@ -51,7 +51,7 @@ build: require-gopath
 
 update-qri-deps: require-gopath
 	cd $$GOPATH/src/github.com/qri-io/qri && git checkout master && git pull && gx install
-	cd $$GOPATH/src/github.com/qri-io/qfs/cafs && git checkout master && git pull
+	cd $$GOPATH/src/github.com/qri-io/qfs && git checkout master && git pull
 	cd $$GOPATH/src/github.com/qri-io/dataset && git checkout master && git pull
 	cd $$GOPATH/src/github.com/qri-io/varName && git checkout master && git pull
 	cd $$GOPATH/src/github.com/qri-io/dsdiff && git checkout master && git pull