--- a/actions/body.go
+++ b/actions/body.go
@@ -3,17 +3,17 @@ package actions
 import (
 	"fmt"
 
-	"github.com/qri-io/cafs"
 	"github.com/qri-io/dataset"
 	"github.com/qri-io/dataset/dsfs"
 	"github.com/qri-io/dataset/dsio"
+	"github.com/qri-io/fs"
 	"github.com/qri-io/qri/p2p"
 )
 
 // LookupBody grabs a subset of a dataset's body
 func LookupBody(node *p2p.QriNode, ds *dataset.Dataset, format dataset.DataFormat, fcfg dataset.FormatConfig, limit, offset int, all bool) (bodyPath string, data []byte, err error) {
 	var (
-		file  cafs.File
+		file  fs.File
 		store = node.Repo.Store()
 	)
 
@@ -28,11 +28,14 @@ func LookupBody(node *p2p.QriNode, ds *dataset.Dataset, format dataset.DataForma
 	}
 
 	st := &dataset.Structure{}
-	st.Assign(ds.Structure, &dataset.Structure{
-		Format:       format,
-		FormatConfig: fcfg,
-		Schema:       ds.Structure.Schema,
-	})
+	assign := &dataset.Structure{
+		Format: format.String(),
+		Schema: ds.Structure.Schema,
+	}
+	if fcfg != nil {
+		assign.FormatConfig = fcfg.Map()
+	}
+	st.Assign(ds.Structure, assign)
 
 	data, err = ConvertBodyFile(file, ds.Structure, st, limit, offset, all)
 	if err != nil {
@@ -45,7 +48,7 @@ func LookupBody(node *p2p.QriNode, ds *dataset.Dataset, format dataset.DataForma
 
 // ConvertBodyFile takes an input file & structure, and converts a specified selection
 // to the structure specified by out
-func ConvertBodyFile(file cafs.File, in, out *dataset.Structure, limit, offset int, all bool) (data []byte, err error) {
+func ConvertBodyFile(file fs.File, in, out *dataset.Structure, limit, offset int, all bool) (data []byte, err error) {
 	buf, err := dsio.NewEntryBuffer(out)
 	if err != nil {
 		err = fmt.Errorf("error allocating result buffer: %s", err)