--- a/actions/dataset_ref.go
+++ b/actions/dataset_ref.go
@@ -48,8 +48,9 @@ func ResolveDatasetRef(node *p2p.QriNode, ref *repo.DatasetRef) (local bool, err
 				responses <- res
 			}()
 
-			var ds *registry.Dataset
-			if ds, res.Error = rc.GetDataset(ref.Peername, ref.Name, ref.ProfileID.String(), ref.Path); res.Error == nil {
+			var rds *registry.Dataset
+			if rds, res.Error = rc.GetDataset(ref.Peername, ref.Name, ref.ProfileID.String(), ref.Path); res.Error == nil {
+				ds := rds.Dataset
 				// Commit author is required to resolve ref
 				if ds.Commit != nil && ds.Commit.Author != nil {
 					ref.Peername = ds.Peername