--- a/actions/dataset_ref.go
+++ b/actions/dataset_ref.go
@@ -50,13 +50,12 @@ func ResolveDatasetRef(node *p2p.QriNode, ref *repo.DatasetRef) (local bool, err
 
 			var rds *registry.Dataset
 			if rds, res.Error = rc.GetDataset(ref.Peername, ref.Name, ref.ProfileID.String(), ref.Path); res.Error == nil {
-				ds := rds.Dataset
 				// Commit author is required to resolve ref
-				if ds.Commit != nil && ds.Commit.Author != nil {
-					ref.Peername = ds.Peername
-					ref.Name = ds.Name
-					ref.ProfileID, _ = profile.IDB58Decode(ds.Commit.Author.ID)
-					ref.Path = ds.Path
+				if rds.Commit != nil && rds.Commit.Author != nil {
+					ref.Peername = rds.Handle
+					ref.Name = rds.Name
+					ref.ProfileID, _ = profile.IDB58Decode(rds.Commit.Author.ID)
+					ref.Path = rds.Path
 					return
 				}
 			}