We're time crunch for a number of qri demos this week and need to ship both frontend and backend releases. Backend has to come first, and this PR contains code we need for demo purposes. So for that reason I'm going to break protocol & merge this today. Despite merging, _this code still requires review from both @dustmop & @ramfox_. I'm hoping you'll both take a read through next week.

Overall, I think you'll like what you see. This PR ships features & has a negative diff stat. What I'm most excited about is all the refactoring that this PR reveals now that many codepaths are easier to reason about. After all this refactoring I think we can kill pervasive structs like `repo.DatasetRef` entirely, and start to draw clearer definitions around qri subsystems like naming, storage, inference & derivation.

The biggest new introduction is qfs ("query filesystem"), which is a new abstraction for working with filesystems. I've tried to define qfs as a _bare minimum assembly of code that was already in our codebase_. We'll work out the roadmap for qfs in a forthcoming RFC.

Given the size of this change, I took time to work through all of our commands, and have tested this branch with all major features of the frontend. Can confirm this branch is less buggy than any previous incarnation. I'm also sure I've missed something, but that's what fixes are for.