@b5 2019-02-03
we don't have proper benchmarks, but speed of test suite execution is a little quicker:

before:
```
ok  	github.com/qri-io/qri/actions	0.460s	coverage: 54.1% of statements
ok  	github.com/qri-io/qri/api	0.418s	coverage: 57.5% of statements
ok  	github.com/qri-io/qri/base	0.077s	coverage: 74.9% of statements
ok  	github.com/qri-io/qri/cmd	16.230s	coverage: 55.4% of statements
ok  	github.com/qri-io/qri/config	0.034s	coverage: 81.8% of statements
ok  	github.com/qri-io/qri/lib	1.098s	coverage: 50.6% of statements
ok  	github.com/qri-io/qri/p2p	1.337s	coverage: 68.6% of statements
ok  	github.com/qri-io/qri/p2p/test	0.159s	coverage: 67.9% of statements
ok  	github.com/qri-io/qri/repo	0.026s	coverage: 72.4% of statements
ok  	github.com/qri-io/qri/repo/fs	0.053s	coverage: 48.5% of statements
ok  	github.com/qri-io/qri/repo/profile	0.023s	coverage: 30.6% of statements
ok  	github.com/qri-io/qri/repo/test	0.054s	coverage: 54.8% of statements
ok  	github.com/qri-io/qri/rev	0.002s	coverage: 93.8% of statements
```
after:
```
ok  	github.com/qri-io/qri/actions	0.402s	coverage: 50.4% of statements
ok  	github.com/qri-io/qri/api	0.382s	coverage: 57.8% of statements
ok  	github.com/qri-io/qri/base	0.071s	coverage: 74.6% of statements
ok  	github.com/qri-io/qri/cmd	14.443s	coverage: 55.6% of statements
ok  	github.com/qri-io/qri/config	0.034s	coverage: 81.8% of statements
ok  	github.com/qri-io/qri/lib	0.907s	coverage: 51.1% of statements
ok  	github.com/qri-io/qri/p2p	1.217s	coverage: 68.6% of statements
ok  	github.com/qri-io/qri/p2p/test	0.147s	coverage: 67.9% of statements
ok  	github.com/qri-io/qri/repo	0.028s	coverage: 72.4% of statements
ok  	github.com/qri-io/qri/repo/fs	0.041s	coverage: 48.4% of statements
ok  	github.com/qri-io/qri/repo/profile	0.023s	coverage: 30.6% of statements
ok  	github.com/qri-io/qri/repo/test	0.057s	coverage: 55.0% of statements
ok  	github.com/qri-io/qri/rev	0.001s	coverage: 93.8% of statements
```

That 2 second gain in cmd tests is nice. circleci test stage is going from ~54sec -> ~49sec.